requirements.txt
fastapi==0.127.0
uvicorn==0.40.0
python-multipart==0.0.21
faster-whisper==1.2.1
pydantic==2.12.5
pydantic[email]==2.12.5
SQLAlchemy==2.0.45
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.5.0
rendercv[full]>=2.6.0
psycopg2-binary>=2.9.11
python-dotenv>=1.2.1
hvac==2.4.0
aiosmtplib==5.1.0
email-validator==2.3.0
Jinja2==3.1.6
httpx==0.28.1